I SOLVED IT
FINALLY I managed to sync the circles to the music. I feed directly the duration of the circle to the spawner and i calculate it outside. Since the spawner is called when there's less than 1 beat left to the next one to be played, I only had to wait for that condition (songpositioninbeats >= nextBeat-1) and then pass as duration of the next circle ((nextBeat-songPositionInBeats) (which is always less than 1) * _secsperBeat).